Move GcsUtil TextIO Tests to TextIOTest
These tests are not a test of the DataflowRunner, nor any
DataflowRunner specific behavior, so they should be part of TextIOTest
